refactor cac keymanager test utility class for package scope access
